bye bye discord.js (start porting this to detritusjs)

+++ events/guildCreate/dequeueDeleteGuildSettings.js
@@ -5,11 +5,15 @@ See tasks/deleteGuildSettings.js for the idea behind this file
  */
 
 module.exports = {
-  fn: (main, guild) => {
+  fn: (main, GuildCreate) => {
+    if (GuildCreate.fromUnavailable) {
+      return;
+    }
+
     main.prometheusMetrics.sqlCommands.labels('DELETE').inc();
     main.db.delete_guild_settings_queue.destroy({
       where: {
-        guild_id: guild.id,
+        guild_id: GuildCreate.guild.id,
       },
     });
   },
